Moved 'stage' types to root of routes folder
